Add expansion, inheritance, include tests (#4137)
* Move the presets tests to integration tests.

These tests aren't true 'unit' tests by definition, so moving them to
integration.
Also, added include and CMakePresets/CMakeUserPresets tests.

* add inheritance and expansion and override tests

* fix formatting of test file

* fix linter issues

* fix lint issues

* fix path

* make tests better across platform

* fix lint
