+++ .github/workflows/ci-main-linux.yml
@@ -56,6 +56,11 @@ jobs:
         with:
           run: yarn unitTests
 
+      - name: Run integration tests
+        uses: GabrielBB/xvfb-action@v1.6
+        with:
+          run: yarn integrationTests
+
       - name: Run successful-build test
         uses: GabrielBB/xvfb-action@v1.6
         with:
