+++ .github/workflows/ci-main.win.yml
@@ -54,6 +54,9 @@ jobs:
       - name: Run unit tests
         run: yarn unitTests
 
+      - name: Run integration tests
+        run: yarn integrationTests
+
       - name: Run successful-build test
         run: yarn endToEndTestsSuccessfulBuild
